fix toss test api에 맞춰서 형변환 (#19)

+++ hotel_project/src/main/java/com/hotel_project/hotel_jpa/room_pricing/dto/IRoomPricing.java
@@ -21,8 +21,8 @@ public interface IRoomPricing extends IId {
     LocalDate getDate();
     void setDate(LocalDate date);
 
-    BigDecimal getPrice();
-    void setPrice(BigDecimal price);
+    Long getPrice();
+    void setPrice(Long price);
 
     LocalDateTime getCreatedAt();
     void setCreatedAt(LocalDateTime createdAt);
